refactor engine.go
